Add AlertManager for congratulations
